\name{convexhull}
\alias{convexhull}
\title{ Convex Hull }
\description{
  Calculates the area and coordinates (for plotting) of the convex hull surrounding bivariate data
}
\usage{
convexhull(x,y)
}
\arguments{
  \item{x,y}{ Bivariate data given as vectors x and y }
}
\author{ Andrew Jackson }
\keyword{ programming }


