\name{correctionsdemo}
\alias{correctionsdemo}
\docType{data}
\title{ Fractionation correction values for the geese data}
\description{
  A 3 column, 2 row matrix containing the mean and standard deviation
  of the correction values for each of the 2 isotopes used
}
\usage{data(correctionsdemo)}
\format{
  A data frame with 2 observations on the following 3 variables.
  \describe{
    \item{\code{Isotope}}{Name of the isotope}
    \item{\code{Mean}}{The mean correction value}
    \item{\code{Sd}}{The standard deviation correction value}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
#see siarmenu() and option 9 for a demo using this data
}
\keyword{datasets}
