\name{geese1demo}
\alias{geese1demo}
\docType{data}
\title{ A single group of the geese data }
\description{
  A 2 column, 9 row matrix containing the plasma data for the first group of geese 
}
\usage{data(geese1demo)}
\format{
  A data frame with 9 observations on the following 2 variables.
  \describe{
    \item{\code{d15NPl}}{d15N plasma}
    \item{\code{d13CPl}}{d13C plasma}
  }
}
\examples{
#see siarmenu() and option 9 for a demo using this data
}
\keyword{datasets}
