\name{geese2demo}
\alias{geese2demo}
\docType{data}
\title{ A multi-group version of the geese plasma data}
\description{
  A 3 column, 251 row matrix which contains the isotopic plasma values of 251 geese over 2 isotopes
}
\usage{data(geese2demo)}
\format{
  A data frame with 251 observations on the following 3 variables.
  \describe{
    \item{\code{Group}}{The group number / time point}
    \item{\code{d15NPl}}{The d15N plasma value}
    \item{\code{d13CPl}}{The d13C plasma value}
  }
}
\examples{
#see siarmenu() and option 9 for a demo using similar data
}
\keyword{datasets}
