\name{hullarea}
\alias{hullarea}
\title{ Convex Hull Area }
\description{
  Utility function to calculate the area enclosed by the convex hull of
  bivariate data. Not intended for direct use by the user, and is called by
  \code{\link{hullarea}}
}
\usage{
hullarea(x,y)
}
\arguments{
  \item{x,y}{ Bivariate data given as vectors x and y }
}
\author{ Andrew Jackson }
\keyword{ programming }