\name{loadsiardata}
\alias{loadsiardata}
\title{ Loads in siar data }
\description{
    Loads in siar data
}
\usage{
loadsiardata(siarversion)
}
\arguments{
  \item{siarversion}{ \item{siarversion}{ The siar version number as a string. } }
}
\details{
  Not intended for use outside \code{\link{siarmenu}}
}
\author{ Andrew Parnell }
\keyword{ programming}
