\name{panelcontour}
\alias{panelcontour}
\title{ Adds contours to a matrix plot }
\description{
  A simple function used by pairs to produce neat looking matrix plots. Not intended for use by
  those using siar.
}
\usage{
panelcontour(x, y, ...)
}
\arguments{
  \item{x}{ A numeric vector containing data with which to produce a contour plot }
  \item{y}{ A numeric vector containing data with which to produce a contour plot }
  \item{\dots}{ Other arguments }
}
\author{ Andrew Parnell }
\keyword{ programming}
\keyword{ models}

