\name{panelcor}
\alias{panelcor}
\title{ Adds correlations to a matrix plot }
\description{
  A simple function used by pairs to produce neat looking matrix plots. Not intended for use by
  those using siar.
}
\usage{
panelcor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
  \item{x}{ A numeric vector containing data with which to produce correlations }
  \item{y}{ A numeric vector containing data with which to produce correlations }
  \item{digits}{ Number of digits to display on plot }
  \item{prefix}{ Text to add before the correlation }
  \item{cex.cor}{ Multiplier for the size of the text on the plot }
  \item{\dots}{ Other arguments }
}
\author{ Unknown }
\keyword{ programming}
\keyword{ models}

