\name{panelhist}
\alias{panelhist}
\title{ Adds histograms to the diagonal of a matrix plot }
\description{
  A simple function used by pairs to produce neat looking matrix plots. Not intended for use by
  those using siar.
}
\usage{
panelhist(x, ...)
}
\arguments{
  \item{x}{ A numeric vector containing data over which to compute a histogram }
  \item{\dots}{ Other arguments }
}
\author{ Uknown }
\keyword{ programming }
