\name{siaraddcross}
\alias{siaraddcross}
\title{ Plotting tool for adding isotope bi-plot data to a figure }
\description{
    A sub-function for siarplotdata() and not intended for calling directly by the user.
}
\author{ Andrew Parnell and Andrew Jackson}
\keyword{ programming}
\keyword{ models}
