\name{siardemo}
\alias{siardemo}
\title{ Runs the siar model and some nice plots for the siar package }
\description{
  A simple function which utilises the loaded in Geese plasma data to run the MCMC on dietary proportions. Can
  be accessed either directly or through the menu function
}
\usage{
siardemo(siarversion = 0)
}
\arguments{
  \item{siarversion}{ Not required }
}
\author{ Andrew Parnell }
\keyword{ programming}
\keyword{ models}
