\name{siarloaddata}
\alias{siarloaddata}
\title{ Loads in siar data }
\description{
    Loads in siar data via a neat menu-driven interface.
}
\usage{
siarloaddata(siarversion)
}
\arguments{
  \item{siarversion}{ The siar version number as a string. }
}
\details{
  Not intended for use outside \code{\link{siarmenu}}
}
\author{ Andrew Parnell }
\keyword{ programming}
