\name{siarmatrixplot}
\alias{siarmatrixplot}
\title{ Matrix plots of siar output }
\description{
    Produces matrix plots of siar output
}
\usage{
siarmatrixplot(siardata, siarversion=0)
}
\arguments{
  \item{siardata}{ A list containing some or all of the following parts: targets, sources, corrections,
  PATH, TITLE, numgroups, numdata, numsources, numiso, SHOULDRUN, GRAPHSONLY, EXIT, and output. For more details
  of these inputs see the \code{\link{siarmenu}} function. }
  \item{siarversion}{ The siar version number.  }
}
\details{
  Not intended for use outside \code{\link{siarmenu}}
}
\author{ Andrew Parnell }
\keyword{ programming}
\keyword{ models}
