\name{siarmenu}
\alias{siarmenu}
\title{A list of menu options for running the siar package}
\description{
Brings up a list of menu options which allow the user to run MCMC and produce some plots
}
\usage{
siarmenu()
}

\value{
}
\references{ }
\author{ Andrew Parnell }
\note{
}
\seealso{ \code{\link{siarmcmc}} }
\examples{
#siarmenu()
}
\keyword{ programming }
\keyword{ models }
\keyword{ multivariate }
