\name{siarplotdatawrapper}
\alias{siarplotdatawrapper}
\title{ Handles repeated plotting instructions for siarplotdata() }
\description{
    A sub-function for siarplotdata() and not intended for calling directly by the user.
}
\author{ Andrew Parnell and Andrew Jackson}
\keyword{ programming}
\keyword{ models}
