\name{siarsaveoutput}
\alias{siarsaveoutput}
\title{ Saves siar output to a file }
\description{
    Saves created siar output to a file
}
\usage{
siarsaveoutput(siardata)
}
\arguments{
  \item{siardata}{ A list containing some or all of the following parts: targets, sources, corrections,
  PATH, TITLE, numgroups, numdata, numsources, numiso, SHOULDRUN, GRAPHSONLY, EXIT, and output. For more details
  of these inputs see the \code{\link{siarmenu}} function. }
}
\details{
  Not intended for use outside \code{\link{siarmenu}}
}
\keyword{ programming}
