\name{sourcesdemo}
\alias{sourcesdemo}
\docType{data}
\title{ Source (in this case plant) isotope values }
\description{
  A 3 column, 4 row matrix containing 4 different plants and their measurements on 2 different isotopes
}
\usage{data(sourcesdemo)}
\format{
  A data frame with 5 observations on the following 3 variables.
  \describe{
    \item{\code{Sources}}{The plants name}
    \item{\code{Meand15N}}{d15N mean}
    \item{\code{SDd15N}}{d15N standard deviation}
    \item{\code{Meand13C}}{d13C mean}
    \item{\code{SDd13C}}{d13C standard deviation}    
  }
}
\examples{
#see siarmenu() and option 9 for a demo using this data
}
\keyword{datasets}
