\name{correctionsdemo}
\alias{correctionsdemo}
\docType{data}
\title{ Fractionation correction values for the geese data}
\description{
  A 5 column, 4 row matrix containing the mean and standard deviation
  of the correction values for each of the 2 isotopes used for each different source
}
\usage{data(correctionsdemo)}
\format{
  A data frame with 4 observations on the following 5 variables.
  \describe{
    \item{\code{Source}}{The source name. }
    \item{\code{Mean15N}}{The mean correction value for 15N}
    \item{\code{Sd15N}}{The standard deviation correction value for 15N}
    \item{\code{Mean13C}}{The mean correction value for 13C}
    \item{\code{Sd13C}}{The standard deviation correction value for 13C}
  }
}
\examples{
#see siarmenu() and option 9 for a demo using this data
}
\keyword{datasets}
